modules/simple_voice_recorder.py
"""
Упрощенный модуль для записи голоса в Discord голосовых каналах
"""
import asyncio
import logging
import wave
import io
import os
from datetime import datetime
from typing import Optional, Dict, Set
import discord
import speech_recognition as sr
import pydub
from pydub import AudioSegment

logger = logging.getLogger(__name__)

class SimpleVoiceRecorder:
    def __init__(self, config, speech_recognizer):
        self.config = config
        self.speech_recognizer = speech_recognizer
        self.recording_active: Dict[int, bool] = {}  # {guild_id: is_recording}
        self.recognizer = sr.Recognizer()
        
        # Настройка распознавателя речи
        self.recognizer.energy_threshold = 300
        self.recognizer.dynamic_energy_threshold = True
        self.recognizer.pause_threshold = 0.8
        self.recognizer.operation_timeout = None
        self.recognizer.phrase_threshold = 0.3
        self.recognizer.non_speaking_duration = 0.5
    
    async def start_recording(self, voice_client: discord.VoiceClient, guild_id: int):
        """Начинает запись голоса в голосовом канале"""
        try:
            if guild_id in self.recording_active and self.recording_active[guild_id]:
                logger.warning(f"Запись уже активна на сервере {guild_id}")
                return False
            
            # В discord.py нет прямого способа записи голоса
            # Это заглушка для демонстрации функциональности
            self.recording_active[guild_id] = True
            logger.info(f"Запись голоса активирована на сервере {guild_id}")
            
            # Запускаем мониторинг голосового канала
            asyncio.create_task(self.monitor_voice_channel(voice_client, guild_id))
            
            return True
            
        except Exception as e:
            logger.error(f"Ошибка начала записи голоса: {e}")
            return False
    
    async def stop_recording(self, guild_id: int):
        """Останавливает запись голоса"""
        try:
            if guild_id in self.recording_active:
                self.recording_active[guild_id] = False
                logger.info(f"Запись голоса остановлена на сервере {guild_id}")
                return True
            return False
        except Exception as e:
            logger.error(f"Ошибка остановки записи голоса: {e}")
            return False
    
    async def monitor_voice_channel(self, voice_client: discord.VoiceClient, guild_id: int):
        """Мониторит голосовой канал (заглушка)"""
        try:
            while self.recording_active.get(guild_id, False):
                # В реальной реализации здесь был бы код для записи аудио
                # Но discord.py не предоставляет прямого доступа к аудио потоку
                
                # Симулируем получение аудио данных
                await asyncio.sleep(1)
                
                # Проверяем, что запись все еще активна
                if not self.recording_active.get(guild_id, False):
                    break
                    
        except Exception as e:
            logger.error(f"Ошибка мониторинга голосового канала: {e}")
        finally:
            self.recording_active[guild_id] = False
    
    async def simulate_voice_recognition(self, guild_id: int, user_id: int, text: str):
        """Симулирует распознавание речи (для демонстрации)"""
        try:
            if not self.config.speech_recognition:
                return
            
            # Проверяем, не исключен ли пользователь
            if self.config.is_user_excluded(guild_id, user_id):
                return
            
            # Логируем симулированную речь
            await self.speech_recognizer.log_speech(
                guild_id, user_id, text, datetime.utcnow()
            )
            
            logger.info(f"Симулированная речь от пользователя {user_id} на сервере {guild_id}: {text}")
            
        except Exception as e:
            logger.error(f"Ошибка симуляции распознавания речи: {e}")
    
    def is_recording(self, guild_id: int) -> bool:
        """Проверяет, идет ли запись на сервере"""
        return self.recording_active.get(guild_id, False)
    
    async def cleanup(self):
        """Очищает все активные записи"""
        for guild_id in list(self.recording_active.keys()):
            await self.stop_recording(guild_id)
