requirements.txt
# Discord Bot Framework (используем discord.py с voice поддержкой)
discord.py[voice]
python-dotenv

# Voice Support
PyNaCl
ffmpeg-python

# Speech Recognition
SpeechRecognition
pydub
pyaudio

# Audio Processing
librosa
soundfile

# Additional Dependencies
aiohttp
aiofiles
numpy
